Update Gradle wrapper script and dependencies
This update modifies the Gradle wrapper script to be more robust and makes several changes to dependencies. Specifically, it updates the version of the Gradle wrapper, the serialization and deserialization classes for multiple formats, and various plugins. It also revises the multi-event consumer to adapt to changes in the serializer classes. Furthermore, it includes some small updates in the handling of .proto files and other minor codes.